Revert "arm64: dts: add NXP i.MX91 device tree"
It is not allowed to add files to dts/upstream that are not a
cherry-pick from the devicetree-rebasing repository. While our
documentation does not spell out that it must be done with
'tools/update-subtree.sh dts pick' it is clear that it must be
cherry-picks and not changes that have not yet been merged.

This reverts commit c4ee9df4a421852fa9fd0376dd4814beb16a1b70.

Signed-off-by: Tom Rini <trini@konsulko.com>
---
Sigh. So here's the proof that not every commit to dts/upstream has been
correct, in mainline. So I need to figure out some better tooling to
enforce using update-subtree.sh pick it seems.

Cc: Peng Fan <peng.fan@nxp.com>
Cc: Sumit Garg <sumit.garg@kernel.org>
Cc: Quentin Schulz <quentin.schulz@cherry.de>